test: migrate webhook-automations to bdd
